# apispec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API DOCUMENTATION",
    "description": "API DOCUMENTATION",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api"
    }
  ],
  "paths": {
    "/register" : {
      "post": {
        "tags": ["Authentication API"],
        "description": "Login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Success",
            "content": {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code" : {
                      "type": "number"
                    },
                    "status" : {
                      "type": "string"
                    },
                    "data" : {
                      "$ref": "#/components/schemas/LoggedInOrRegistered"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Register Failed",
            "content": {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code" : {
                      "type": "number"
                    },
                    "status" : {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "security": [{
          "JWT" : []
        }],
        "tags": ["Authentication API"],
        "description": "Login",
        "requestBody": {
          "content": {
            "application/json" : {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login Success",
            "content": {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code" : {
                      "type": "number"
                    },
                    "status" : {
                      "type": "string"
                    },
                    "data" : {
                      "$ref": "#/components/schemas/LoggedInOrRegistered"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Login Failed",
            "content": {
              "application/json" : {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code" : {
                      "type": "number"
                    },
                    "status" : {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "JWT" : {
        "type": "apiKey",
        "in": "header",
        "name": "API-KEY",
        "description": "Authentication for API"
      }
    },
    "schemas": {
      "Login": {
        "type": "object",
        "properties": {
          "username" : {
            "type": "string",
            "required": true
          },
          "password" : {
            "type": "string",
            "required": true
          }
        }
      },
      "Register" : {
        "type": "object",
        "properties": {
          "username" : {
            "type": "string",
            "required": true
          },
          "email" : {
            "type": "string",
            "required": true
          },
          "password" : {
            "type": "string",
            "required": true
          }
        }
      },
      "LoggedInOrRegistered" : {
        "type": "object",
        "properties": {
          "id" : {
            "type": "number"
          },
          "username" : {
            "type": "string"
          },
          "token" : {
            "type": "string"
          }
        }
      },
...
    }
  }
}